+++ lib/contentful_model/errors.rb
@@ -1,5 +1,4 @@
 module ContentfulModel
   class AssociationError < StandardError; end
   class VersionMismatchError < StandardError; end
-  class AttributeNotFoundError < NoMethodError; end
 end