+++ lib/contentful_model/management.rb
@@ -1,4 +1,5 @@
 module ContentfulModel
+  # Wrapper for the CMA Client
   class Management < Contentful::Management::Client
     def initialize(options = {})
       super(ContentfulModel.configuration.management_token, options)